Fixed some armv7/32-bit pf issues

+++ lib/allocation/garbage-collection.c
@@ -256,11 +256,11 @@ void lGarbageCollect(){
 	if(lVerbose){
 		const u64 end = getMSecs();
 		pf("== Garbage Collection #%u took %ims ==\n",lGCRuns,(i64)(end-start));
-		pf("Vals: %u -> %u [Δ %i]{Σ %i}\n",bva,lValActive,    (i64)lValActive     - bva, lValMax);
-		pf("Clos: %u -> %u [Δ %i]{Σ %i}\n",bca,lClosureActive,(i64)lClosureActive - bca, lClosureMax);
-		pf("Strs: %u -> %u [Δ %i]{Σ %i}\n",bsa,lStringActive, (i64)lStringActive  - bsa, lStringMax);
-		pf("Arrs: %u -> %u [Δ %i]{Σ %i}\n",baa,lArrayActive,  (i64)lArrayActive   - baa, lArrayMax);
-		pf("Tres: %u -> %u [Δ %i]{Σ %i}\n",bta,lTreeActive,   (i64)lTreeActive    - bta, lTreeMax);
+		pf("Vals: %u -> %u [Δ %i]{Σ %i}\n",(i64)bva,(i64)lValActive,    (i64)lValActive     - bva, (i64)lValMax);
+		pf("Clos: %u -> %u [Δ %i]{Σ %i}\n",(i64)bca,(i64)lClosureActive,(i64)lClosureActive - bca, (i64)lClosureMax);
+		pf("Strs: %u -> %u [Δ %i]{Σ %i}\n",(i64)bsa,(i64)lStringActive, (i64)lStringActive  - bsa, (i64)lStringMax);
+		pf("Arrs: %u -> %u [Δ %i]{Σ %i}\n",(i64)baa,(i64)lArrayActive,  (i64)lArrayActive   - baa, (i64)lArrayMax);
+		pf("Tres: %u -> %u [Δ %i]{Σ %i}\n",(i64)bta,(i64)lTreeActive,   (i64)lTreeActive    - bta, (i64)lTreeMax);
 		pf("--------------\n\n");
 	}
         lCheckBreak();
